cloud-run-job/pyproject.toml
[tool.poetry]
name = "run-rag"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
pydantic = "^2.1"
unstructured = {extras = ["md", "pdf"], version = "^0.16.17"}
langchain-community = "0.3.14"
pgvector = "^0.3.6"
langchain-google-cloud-sql-pg = "0.12.0"
langchain-google-vertexai = "^2.0.12"
torch = { version = "^2.6.0", source = "pypi", markers = "sys_platform != 'win32'" }
google-cloud-storage = "2.19.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
